requirements.txt
openai==1.30.3
Flask==3.0.0
Flask-Cors==4.0.0
langchain==0.2.1
langchainhub==0.1.14
langchain-openai==0.1.7
langchain-community==0.2.1
chromadb==0.5.0
langchain-chroma==0.1.1
langgraph==0.0.60
streamlit==1.35.0
tabulate==0.9.0
streamlit-mic-recorder==0.0.8
python-dotenv==1.0.0